extracts avatar to own component
